+++ Sources/DeckKit/_Deprecated/Deprecated.swift
@@ -1,0 +1,2 @@
+@available(*, deprecated, message: "Just use Identifiable instead.")
+public protocol DeckItem: Identifiable {}